+++ src/features/notifications/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useNotifications } from './useNotifications';
